BAU: ensure we do not wipe sms details when setting mfa methods migrated to false
When we set mfa methods migrated, we were wiping the sms details from the user profile table regardless of whether we were setting migrated to true or false. It makes sense when setting to true, as we will have transferred this information over to the user credentials table.

When setting explicitly to false however, we should not be wiping this information.

We only do this in tests currently (for live users this field is empty), so this was not causing a live issue, but just unexpected behaviour in tests.