AUT-1416: Convert from `MFAMethod` to `MfaMethodResponse`
The `MfaMethodsService` returns an API agnostic record instances `MFAMethod`, which need to be converted to the new `MfaMethodResponse` for frontend-api responses.

This new converter takes a set of `MFAMethod`s and utilising a new static `from` method on `MfaMethodResponse` does the conversion into the correct record types.

This is in anticipation of needing the `LoginHandler` to return the `mfaMethods` of the user in the new `MfaMethodResponse` record.