AUT-1416: Create Gson `TypeAdaptor` for `MfaMethodResponse`
We will soon be updating the `LoginResponse` to return a `List<MfaMethodResponse` so the frontend can know all active `mfaMethods` after the user submits the correct password.

We have test of the `LoginHandler` right now that deserialize the response and checks it against an expected `LoginResponse`.

Gson does not know how to deserialise the `mfaMethods` array as items in there could be converted into a `SmsMfaMethodResponse` or `AuthAppMfaMethodResponse` depending on the `type` attribute.

Here we introduce the `MfaMethodResponseAdaptor` to do this conversion.

We also modify `SerializationService` to allow `extraTypeAdapters` to be passed in. This is done because the `SerializationService` is shared across APIs and we do not want to leak `frontend-api` specific implementations of `MfaMethodResponse` to other APIs that have a different implementation (i.e. account-management).