ATO-1566: Add issuer to allowed aud values
We plan to allow the issuer value to be used as the audience
value when using private_key_jwt authentication method. This
adds an addtional value we permit.

Note that adding this means we now support providing both
audience values in the private_key_jwt and have added a test
documenting this behaviour