AUT-4202: Introduce specific MfaMethodUpdateRequest class
The MfaMethodCreateOrUpdateRequest was intended to be generic for both
create (POST) and update (PUT) requests.  This differed from the OpenAPI
specification which has a distinct schema for each to allow for the
swap feature where a BACKUP MFA method can be made the DEFAULT.

A new class specifically for method updates is needed that allows the
'method' field to be omitted for a swap request.