AUT-4202: Enforce business rules for BACKUP MFA methods
A BACKUP MFA method cannot be updated.  The only action a User can take
is to swap the BACKUP for the DEFAULT.

Remove the checks for details of the update request as they should not
be sent.  Introduce a new error code stating that a BACKUP method cannot
be edited (as opposed to updated which is allowed).

Testing focus is now on the swap feature.