BAU: Add decorator to client sessions list
In previous work we did not add the usual DynamoDbAttribute decorator
to the getter for this value. This meant it defaulted to using the
the attribute name `clientSessions` instead of `ClientSessions` which
was defined as a variable.

Here we add this decorator but change the attribute to match what
the current value is in Dynamo. This is required for the changes to
be backwards compatiable with existing session fields. Doing this
now also prevents someone from doing this in the future with the
exsiting attribute string, which would be breaking