ATO-1621: Set achieved credential strengh on SMS sign in
We currently don't set the current credential strength in this handler
when a user is on an SMS sign in journey. This has meant that for some journey
types, this value is null where a user has been on a medium journey.

To rectify this, and whilst we migrate to the new field, we should set
the achieved credential trust to a medium level for SMS MFA journeys.