AUT-4309: Fix MFA method switch
We had a business rule that prevented the type of the Default MFA
Method from being changed, i.e. stop changing an SMS to a AUTH APP
and vice versa.  This is not a valid rule and has been removed.

Now the Update MFA service allows the DEFAULT MFA to be updated
correctly and the BACKUP can be updated to cause it to switch with the
DEFAULT.