ATO-1621: Don't downshift credential strength
It is possible that a user can go on a medium followed by a low
credential strength journey. In these cases, we shouldn't down shift
them and they should retain the medium value they previously achieved.
This partially replicates some of the existing behaviour in orchestration which
happens at authentication callback with the shared session[1].

To achieve this, I've added two new methods to the enum which are just wrappers
around the compareTo method. I didn't really like the usage of <0, >0 etc that
the usage of compareTo forces you to do, as it seems unclear what the intended
check is when looking at an if statement. I realise this could probably be written
with a check for just not medium, but these methods leave room for additonal comparison if
we have higher values of credential trust in future

[1]- https://github.com/govuk-one-login/authentication-api/blob/9cfee94852b6bc05463e36f44b1264f7bbefab14/oidc-api/src/main/java/uk/gov/di/authentication/oidc/lambda/AuthenticationCallbackHandler.java#L572