AUT-4248: Generalised OTP key variables
The `CodeStorageService` has the following methods; `saveOtpCode`, `getOtpCode` & `deleteOtpCode`.

These are used in a variety of journies. With support for backup MFAs we need SMS codes to be stored with reference to the phone number in addition to email address.

Therefore the key used is no longer just email addresses, but will also include an phone number.