AUT-4248: Add `isForPhoneNumber` method to `NotificationType`
We need to distinguish between notifications sent to phone numbers vs other destinations like email in order to determine how to store OTP code based on the `NotificationType` sent to the backend lambdas from the frontend.

This adds a method that returns a boolean, similar to the existing `isEmail` method.