AUT-4248: `MfaHandler` SMS OTP code identified by phone number & email address
`MfaHander`s only job is to send SMS OTP codes for a variety of journeys.

We need to have unique OTP records for each phone number used by a user. At the moment they are stored in redis with an ID derived from just the users email address.

Going forward this will not suffice. Users will be able to have multiple SMS OTP methods and we need to ID each numbers own OTP codes.

So here we add the phone number to the email address identifier of OTP codes.

Fetching the OTP codes for verification with these new IDs will be handled in the upcoming commits.