AUT-4248: `SendNotification` SMS OTP code identified by phone number & email address
`SendNotification` is responsible for sending:
- the `VERIFY_EMAIL` & `ACCOUNT_CREATED_CONFIRMATION` notifications to email addresses
- the `VERIFY_PHONE_NUMBER` and `MFA_SMS` notifications to phone numbers

We need to have unique OTP records for each phone number used by a user. At the moment they are stored in redis with an ID derived from just the users email address.

Going forward this will not suffice. Users will be able to have multiple SMS OTP methods and we need to ID each numbers own OTP codes.

So here we add the phone number to the email address identifier of OTP codes only when the `notificationType` is for a phone number.

Fetching the OTP codes for verification with these new IDs will be handled in the upcoming commits.