AUT-4248: `VerifyCodeHandler` SMS OTP code identified by phone number & email address
`VerifyCodeHandler` is responsible for verifying OTPs sent via SMS and email in certain journeys.

We now have unique OTP records stored against each user email address and phone number combo.

So here we ensure that verification is done with the OTP recorded against the unique phone number and email address identifier.

As this API does not accept any identifier for the phone number in use, we have to pull the default SMS for the user and assume this is the one to verify against (as is current behaviour). This API will be modified in the future to accept an optional MFA identifier to help here.