OLH-2621: Check for interventions in authenticate handler
Call AIS to check for interventions in the authenticate handler. If the
user has a blocked or suspended intervention, return a 403 and an error
code.

When a user goes to change eg. their password or MFA settings in the
account management frontend they're prompted to re-enter their password.
That's checked by calling this API route. If the user has an
intervention that should stop them using their account, the account
management frontend needs to be able to block that change and log the
user out to prevent any further changes or access.