AUT-4320: Remove redundant code
The ValidationHelper#validatePhoneNumber method contained
code that could never be reached.  This was highlighted by
SonarQUBE warning that the code coverage for this class was
too low and then trying to find a test phone number that
showed the called to phoneUtil.isValidNumber(phoneNumber)
could fail.

There isn't a phone number that will fail the checks in
the isValidNumber method given the checks already
performed.