ATO-1621: Stop sending current credential strength to Orch
We're now aiming to remove the circular logic which was previously
established during the session migration. The thing this value
represents - the user's level of credential trust - is now
represented with a new value called Achieved Credential Strength.

The code in Orchestration handles this value being null already,
so we can remove this here.