AUT-4320: Move business rules out of persistence class
Move the business rules for updating an MFAMethod out of
the DynamoService and into the business layer class.

This makes the persistence layer easier to test.

The business rules around what values should be set in an
MFAMethod depending on its type are now in MFAMethodsService
with other business rules making it easier to see what the
code is doing.