AUT-4256: ResetPasswordRequest now returns mfaMethods
Our reset password journey needs to know which MFA methods a user has in order to present them as choices after entering their email OTP successfully.

Here we add the `mfaMethods` to the `ResetPasswordRequest` handler, as we do with the LoginHandler.

Existing returned attributes remain the same, but the data source is updated to use the same `MFAMethodsService` response that is needed for getting the `mfaMethods`.

Permission is given to the lamda to write user data as it now needs write generated MFA IDs for MFA methods that don't have them (`MFAMethodsService` handles that specifically).