AUT-4249: Allow VerifyMfaCode API to use BACKUP AUTH_APP
Previously we had restrictive logic that only allowed a AUTH_APP to be verified if it was the DEFAULT MFA method.

Here this is broadened to allow use a users AUTH_APP MFA method that has any priority.

This is OK for now as we only allow one AUTH_APP MFA method per user. Should this change in the future the VerifyMfaCode API will need a new `mfaMethodId` attribute to allow the frontend to specify which MFA method the code is for.