AUT-4320: Move phone in use check
Moving the check that a phone number must not be in use as an
MFA before we send an otp.  This is a business rule that we
only send otps when a new phone number is being added as an
MFA Method.

Tidied tests and added tests for phone in use check.