AUT-4320: Introduce read only mode for MFA retrieval
When the SendOtpNotificationHandler is asked to send
an otp it is because the user wants to add the number
as an SMS MFA method.  If the number is already in use
and has been verified then an otp should not be sent
as the request will not cause any change to the Users
account.

If the User is currently using the phone number as an
MFA but the phone number has not been verified then we
allow them to send an otp as this will verify their
ownership of the phone number.  There are a small
number of Users in this state and they are all
un-migrated Users, i.e. they created their account
before MFA Method Management went live.