AUT-4426: `CodeStorageService` increment MFA uses in single count
Going forward we will be merging the seperate counts for uses of SMS and AUTH_APP MFA into a single count to support simple lockout logic when a user has multiple MFAs.

Here the `CodeStorageService` and their uses are updated to reflect this. It recorded seperate redis records for the counts.

As this change will affect the uses of existing counts, we need to make sure the now deprecated count records are included in lockout logic until the depecated count records expire.

This means the inclusion of temporary code that can be removed later to ensure a zero-downtime deployment with no user facing changes.