ATO-982: Remove max age redis session handling
We are aiming to removing setting the shared redis session in this
handler. This requires us to remove all the cases we handle the shared
session, which means removing it from the max_age expired handling.

Here I'm temp hardcoding this to creating a new session object
to ensure this commit build successfully. This will soon be removed
in a followup commit