AUT-4379: Add MFA details to AUTH_PHONE_CODE_SENT
The AUTH_PHONE_CODE_SENT audit event should have DEFAULT as the MFA
type for non-migrated Users.

For migrated Users add the MFA type for the MFA Method for the
phone number.

If the phone number is not currently used as an MFA then set the
mfa type in the audit event to be DEFAULT.  The justification for
this is that DAP will be able to recognise when this is incorrect
by looking at the flow of events.