AUT-4452: MFAMethodType agnostic CodeStorageService CODE(_REQUEST)_BLOCKED keys
Currently we have locks set in redis using the `CODE_BLOCKED_KEY_PREFIX` and `CODE_REQUEST_BLOCKED_KEY_PREFIX` prefixes. These are then appended with a concatenation of JourneyType and MFAMethodType.

Going forward we want to make this lock MFAMethodType agnostic (so locks triggered by SMS or AUTH_APP are the same lock).

This will have the effect blocking entering an MFA OTP on the users DEFAULT MFA method if they managed to lock themselves out by entering too many incorrect OTPs using their BACKUP on the same journey and where DEFAULT and BACKUP are different MFAMethodTypes. Similar for when a user has locked out for requesting too many SMSs.

This is done by narrowing the possible options in the CodeRequestType enum, generalising SMS and AUTH_APP mentions to a generic MFA value (as an aside concatenation of the CodeRequestType for the PASSWORD_RESET_MFA journey is flipped to match the others).

This also has the effect of ignoring any existing lockouts for the deprecated prefixes. This will be addressed in a follow on commit.