AUT-4452: Check MFA lockout just after password entry
Now that we have merged counts so they are MFAMethodType agnostic, we need to ensure that when a BACKUP SMS MFA usage triggers a lockout, the user is not permitted to restart their journey using their DEFAULT AUTH_APP MFA.

They can do so currently. This is because we do not check after the user has entered their password for any MFA lockouts before redirecting to the AUTH_APP entry page. If their DEFAULT was an SMS MFA, then the API call to send the SMS OTP would return the error instead.

Here we add the MFA check to the LoginHandler. When it finds an MFA block it will return the associated error. The frontend can then handle show the user a relevant error message and prevent redirection to the AUTH_APP OTP entry page.