AUT-4452: Ensure password reset email OTP verification returns error to user if MFA submission is locked out
For a user to start a password reset journey they need to verify access to the email account first. This means an OTP is sent to the email which they then submit to the UI. The VerifyCodeHandler checks the OTP is correct before the user is allowed to move onto the next part, which is verifying access to their MFA.

Previously we would then trigger an SMS OTP sending if the user has an SMS MFA. That handler would error if the user is already locked out of using their MFA.

For Method Management we have merged lockout counts. So when the user has both an SMS and AUTH_APP MFA method, an OTP entry on either increments the same count. This also means lockouts for MFA usage are also MFAMethodType agnostic.

The problem we have is that if a user has an AUTH_APP MFA as their default, and they have previously locked themselves out of password reset by entering too many incorrect OTPs on their BACKUP SMS MFA, they are still redirected to their AUTH_APP OTP entry page after entering their email OTP.

As counts are now MFAMethodType agnostic, this should not be the case. Instead we should show the user an error straight after email OTP entry if they have previously locked themselves out with either their SMS or AUTH_APP MFA.

So here the VerifyCodeHandler is updated to do that check. There is a separate frontend change to respond to the returned errors.