AUT-4315: use 'read only' option for getMfaMethods
The getMfaMethods method actually makes updates to the database as part of the migration process, and this requires write access to user-profile.  As we don't always want this to happen there is an option to skip the updates, which avoids having to grant update rights to calling lambdas.

This applies in this case so using the read only option when getting the methods.