AUT-4372: Refactor AuthenticationSuccess metric helper methods
We use the same helper method when an MFA is used to and when it was not used. This is a little bit confusing when working with the method.

So here this is split out to be more explicit. This is in preparation to when we add more dimensions to the metric, but only when an MFA is used.