AUT-4382: Add `mfa-method` audit extension to `AUTH_CODE_VERIFIED`
When we verify an MFA code we now need to know if it was the users DEFAULT or BACKUP MFA method.

Note that the VerifyCodeHandler is used for both email and SMS code verification, but only SMS is used in the context of an MFA, so it assumed that the data only refers to an SMS MFA.

Note that the VerifyMfaCodeHandler is used for both SMS and AUTH_APP code verification. SMS is only checked in the context of `ACCOUNT_RECOVERY` and `REGISTRATION` where the priority is DEFAULT. AUTH_APP is used for all journey types, so a lookup is done to find the priority of the 1 possible AUTH_APP method on the account (falling back to DEFAULT for `ACCOUNT_RECOVERY` and `REGISTRATION` journeys).