ATO-1780: Add RateLimitDecision class
This class represents the data type that contains our rate limit
decision. Whilst I realise the interface could _currently_ just
return a boolean, I wanted to make this more flexible by creating
an enum of RateLimitActions which represent how we might respond
in the case an RP has exceeded the rate limit incase this is not
a binary choice in the future.

I think this offers a bit more flexibility in testing too, as
we can test different combinations of states and outcomes. I have
disallowed one state in the constructor which is performing an
action when the rate limit is not exceeded - this makes sense to me
given we don't want to allow an action if no rate is limit is exceeded