AUT-4382: Refactor audit event metadata pairs in MfaHandler
In preparation for adding an extra `mfa-method` metadata pair after the first initialisation of `journey-type` and `mfa-type`, the metadata pairs need to be set up in a mutable way.

`auditContext` was recently given a way to do this with the `withMetadataItem` method, allowing the removal of a separate variable just for holding a local `AuditService.MetadataPair[]`.

This allows the simpler addition of the `mfa-method` pair in a follow on commit.