OLH-2696: Create wrapper MfaUpdateFailure
We will soon be adding audit events for when MFA update fails. The current MFAMethodsService update method will return the type of update on success (e.g. if it was a switch update).

As the audit events will be tailored to the type of update, we also need to return that on failure too.

So this commit starts getting ready for this by creating a wrapper. Instead of the update methods returning the MfaUpdateFailureReason on failure, they now return MfaUpdateFailure, which wraps the reason. MfaUpdateFailure will be updated in an upcoming commit to include the update type too.

This will then be used to tailor the failure audit event.