AUT-4512: force api deployment with a document change
The api switch on its own will not cause the api to redeploy, which caused an issue when switching on in staging.  As the switch is temporary and won't be used again it's not worth forcing a deployment when it is changed, so this is a one-off change to make sure the api deploys correctly.

Deployments do take place whenever the api spec changes and lambda changes have been proven to be deployed.  These are the ongoing changes that will need to be supported going forward and they are working correctly.