BAU: Fix MFAMethodsService integration test data
The following commits will be tweaking and adding tests following changes to the MFAMethodsService.

It became apparent that the previous version of this file was allowing individual tests to alter the test data used in subsequent tests in the same file. This passes now as we're not asserting on any of the changed data. This will change in the next commits.

So this is updated to use CommonTestVariables and clone the shared MFA methods before each test to get a unique object for each test.

Then we update assertions on lists of MFA methods, to use the assertIterableEquals in order to utilise the equality check function defined in the MFAMethod. For that function, we do not care about the updated timestamp being the same, so this is removed.