AUT-4526: Create getMfaMethod function
In an upcoming commit we need to fetch a particular MFA method. This is so we can get data about the method for an audit event that needs raising before we update the method.

Here we create a method that will return the method asked for. This introduces a new `MfaRetrieveFailureReason` of `UNKNOWN_MFA_IDENTIFIER`. This also returns the full list of MFAMethods of the user as this will also be required.