AUT-4526: Fetch MFAMethod for use before updating it
In a coming commit an audit event will be raised with data from the MFA method before it is updated. To do this the `updateMfaMethod` is updated to explicitly not do the initial fetch. Uses of the update function now need to provide the selected MFAMethod and all MFAMethods of the user. In a previous commit a new `getMfaMethod` function is created and here it is used.