AUT-4525: Make it explicit where events originate
By using the share AuditService we inherited the default setting
of component id equalling AUTH for all audit events emitted from
the account-management-api.

Added a new constant for AUTH component id, updated all calls to
the AuditService to use the new convenience method that allows
the component id to over-ridden.