AUT-4560: Creating `user-permissions` module
We will be defining a `UserPermissionsService` which will be holding the business logic and data store accesses for determining if a user is allowed to do a thing.

Here a new module is created to completely isolate the new code and enforce the API for the consumers in frontend-api. We do not want to leak any internal workings.