AUT-4560: Create UserPermissionContext
Due to the 3 different lockout data stores in use, several different identifiers are used. When we create service methods for UserPermissions they will need to be passed this context in order to interact with the correct records in existing datastores.

- `internalSubjectId` and `rpPairwiseId` are needed for AuthenticationAttemptsService interactions
- `emailAddress` is needed for CodeStorageService interactions
- `authSessionItem` is needed for AuthSessionService interactions

In future phases we can consolidate the datastores and hopefully reduce the context needed.

Not all context will be needed, depending on the journey and action. Until we are able to reduce the context it is proposed that method implementations return errors if they do not have the necessary context for a datastore interaction.