AUT-4560: Add *Error records
When we fail to make a decision if a user can do an action, or fail to record a users action we should return one of these errors (`UserPermissions` methods will be using the `Result<Failure, Success>` return pattern).

Only `UNKNOWN` is declared at the moment to have a value for the enum. The first implemented failure case should replace this (such as error accessing datastore).