AUT-4560: Add Decision
This will be returned when the UserPermissionsService has determined if a user can do an action or not.

To start, two variants are created.

- `Permitted` indicated the user is allowed to do the action requested
- `TemporarilyLockedOut` indicates the user is not allowed to do the action requested for a short time.

The variants can contain different properties, to help give context to the handlers as how the decision was made and provide useful data (for things like audit events).

All variants must contain `attemptCount` for now. This may change in the future but we almost always want to record the attempt number in an audit event even if the action is permitted.

The variants can grow over time as required. For now, `TemporarilyLockedOut` will cover all the expiring lockout cases we have.