AUT-4560: Add methods for verify MFA OTP actions
Create methods for verifying the MFA OTP and recording when the correct or incorrect MFA is received.

The intent in the UserPermissionService is to handle the business logic for the lockout and work with the right backend datastores to increment, reset and check the counts and blocks in different JourneyType contexts.

This leaves a straightforward API for handlers to use to check and record user actions, without having to know about how it's implemented.

This also adds an example handler to give a demonstration of how this can be used.