ATO-1539: Make the "claim" field a JSON object
The ordering of keys in the JSON string is not deterministic
and therefore this test will fail everytime it is run as the
key ordering in the JSON string will continue to swap each time.

To resolve this I've added a fudge - we reformat the claim field into
a JSON object before passing it to our approval test. This should
hopefully mean this is more stable. This also requires us to specifically
serialize nulls, otherwise the userinfo claims are not included!