AUT-4528: Remove mfa-type from UPDATE_PHONE_NUMBER
Code in the AuditService was adding mfa-type to the extensions by
default but this field is not required for the UPDATE_PHONE_NUMBER
audit event.

The easiest fix is to add a filter to the AuditService to exclude
UPDATE_PHONE_NUMBER from this rather than removing this behaviour
from the AuditService.  Also, we are going to rework how the
audit events are built in the next sprint so this quick fix will not
be in place for long.