INCIDEN-1031: Create DeprecationChecker project
This new project will be used as an automatic tool to check important parts of code are marked deprecated in the `main` branch before they can be removed. The intent is another safety layer for when we are removing things like enum values where doing so may cause an incident without careful consideration of the implications for users with active sessions.