INCIDEN-1031: Add methods to get file content
In order to check flagged java enums are annotated as deprecated in main before removal we need to fetch a list of java files to compare, fetch their content in the workspace and fetch the committed content on main.

Here we create a function to get all java files. A future commit could optimise this a bit.