AUT-4449: Remove unneeded Redis configuration
Redis config is no longer required for these lambdas, so the envars and
security groups related to Redis have been removed.

The policy attachment has been retained in production, as removing it
will likely cause a brief outage while the role is updated. This can be
removed in a future PR.