ATO-1856: Add method to validate correct client session id for error
Adds a method which validates a couple of things are correct:

Checks that if we got an access_denied error with state.
As part of this set of changes we will continue with these
checks if we've been given a cross browser error from IPV
which is denoted with an access_denied error code. We also
need the state to be able to effectively respond to this
so we will have to pass on requests which are missing state
for now

If we have got that error we do some additional checking to assert
that we continue to have the same client session id which is associated
with the state value. This requires us to pull some values from different
data stores - we throw a no session error if the state cannot be found as
it is highly likley we can't continue if this value is not present.

If we have the same client session id from the cookie and the state value
we can continue as normal - hence we return early. Otherwise we know
we're in a mismatch situation - in which case we can fetch the client session
using the identifier from derived from the state value - this will ensure we
are using the correct RP to redirect to - ie the one that initiated the authorisation
request to IPV