ATO-1856: Add feature flagged enhanced cross browser handling
This adds the new enhanced cross browser handling we use added in the
previous commit with the feature flag we added previously. I've added
it here in the handler as we want this to occur just after the cookies
have been parsed and the session has been validated - but before we undertake
any additonal actions. This includes our callback validation - as this uses
the cookie derived client session to return errors back to the RP for invalid
callbacks