BAU: Add notification / session ID reference to notification failure log
The notification reference is auto-generated upon NotifyRequest instantiation and is currently logged within those handlers. The client session id may not be set and so this part of the reference may have a stringified "null" reference - we send this to notify anyway.

This log addition will help us track errors across lambdas more easily. If there are send errors we should be able to associate these with session ids using this reference (as it is logged in multiple lambdas).