ATO-1593: Sort Auth userinfo claims
This makes no difference to how auth parse our requested claims as
JSON is order neutral. However becasue we serialize this as a string
in the approvals tests, the ordering of these claims needs to be
fixed. By sorting them here - we can ensure they remain in the same
order between runs