AUT-4569: Add storage layer error
Previously the CheckUserExistsHandler called the CodeStorageService
directly to check if the user had a lock preventing them from entering
their password.  If the call to CodeStorageService failed then an
escaped exception would result in the lambda returning a 500 error
response (and a cold start for the lambda).

Now we have explicit error checking around the call via the new
PermissionDecisionManager we can improve this by returning a new
ErrorResponse of STORAGE_LAYER_ERROR in a 500 response without
suffering a cold start.