AUT-4287: Rotate MFA reset signing key to secondary on staging and below
Note that on the integration and production environments we should still be using the old/same signing key at this point.

Both signing keys (primary and secondary) will continue being published on the JWKS endpoint on staging and below, with only the primary being published on the higher envs (flagged off).