BAU: send correct signing public key in JWT header
Before we were sending the alias in the headers as the keyID. This
was causing a bug where IPV could not use the kid to grab the key from
our jwks. This commit addresses this by grabbing the public key ID from
KMS, hashing it and adding it to the headers. This is what IPV are
expecting