BAU: Only compare against merge base
When a new PR is open, sometimes the DeprecationChecker will flag an issue if the current branch has not been rebased with main where main has just had a new value added to a protected enum.

Here we make sure the comparison to find removed enums only considers removal from merge base with main.

That way if a new value is added to main, but the branch not rebased it won't flag a violation.