AUT-4287: Publish deprecated alias on staging and below.
Note that as we are performing a staged rollout, we need to switch the env vars such that:
- On staging and below we are publishing two keys: the new key through the main alias, and the current key (that's being deprecated) through the deprecated alias.
- On integration and production we are publishing the same single key and no deprecated key.

Note that we are using the `ipv_reverification_request_signing_key_v1_alias` here instead of `ipv_reverification_request_signing_key_alias`. This is as when we rotate we don't want to be publishing the same key twice on staging and below (which would happen with this change if we didn't include the "v1"). These aliases both have the same target key id, so the same key should continue to be published, worst case IPV will use their hardcoded fallback here (we can see this through the Splunk logs).

This has been gated using `var.environment` (as done elsewhere).