AUT-4640: Sign using new v2 key on production
Note that at this point all flagging around this rotation will be removed. We will be both publishing both the v1 and v2 keys AND signing with the new v2 key on all environments.

Revert strategy (pretty self-explanatory but adding here): should just be reverting this change to the  `ipv_reverification_request_signing_key_alias` target_key_id, can be done through AWS UI if need be. Requests will then be signed with the old (v1) key again, and as we are still publishing this, alongside the new v2 key, subsequent requests using that v1 key should succeed again.