AUT-4640: Stop publishing v1 key on JWKS endpoint on production
Note that at this point we are signing with the v2 key on all environments successfully, and the v1 key is only being published on production still.

Also note that we can safely remove the `IPV_REVERIFICATION_REQUESTS_SIGNING_KEY_DEPRECATED_ALIAS` env var from this JWKS function safely as if there is no value defined then this is handled and no secondary signing key is published.