AUT-4640: Remove v1 MFA reset signing key resource and alias
The v1 alias resource is not referenced anywhere (also searched for references in the CF files and can't see any) and the v1 key resource is only referenced by that alias.

The lambdas now use the v2 signing key for both publishing and signing.

It should be safe to remove both of these resources at the same time, I believe TF should understand the dependency between these.

Note that the alias will be deleted immediately and the key resource should respect the `deletion_window_in_days` attribute that was set on it and be marked for deletion to allow for recovery within this period.